requirements.txt
flask==2.0.1
werkzeug==2.0.1
flask-cors==3.0.10
requests==2.28.2
python-dotenv==0.21.1 